Перенос ошибок спрайтинга в texture_manager
