+++ Code/texture_manager/__init__.py
@@ -1,3 +1,5 @@
-from Code.texture_manager.color.hex_color import HEXColor
-from Code.texture_manager.color.rgb_color import RGBColor
-from Code.texture_manager.texture_manager import TextureManager
+from texture_manager.color.hex_color import HEXColor
+from texture_manager.color.rgb_color import RGBColor
+from texture_manager.texture_errors import (InvalidSpriteError,
+                                            SpriteValidationError)
+from texture_manager.texture_manager import TextureManager
