--- Tests/__init__.py
@@ -1,1 +1,0 @@
-from Tests.Texture.ValidateSprite import InvalidSpriteError, SpriteValidationError
